crypto/config/requirements.txt
# Core dependencies for Trend-Following Trading System
aiohttp>=3.8.0
requests>=2.28.0
web3>=6.0.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
asyncio
typing-extensions>=4.0.0

# Technical Analysis
ta-lib>=0.4.25  # Technical Analysis Library
scipy>=1.9.0

# Machine Learning for Trading
xgboost>=1.7.0          # XGBoost for ML predictions
scikit-learn>=1.2.0     # Machine learning utilities
joblib>=1.2.0          # Model serialization

# Portfolio Optimization & Risk Management
pyportfolioopt>=1.5.0   # Advanced portfolio optimization

# LLM Integration (optional)
openai>=1.0.0

# Logging and utilities
colorama>=0.4.6
python-dateutil>=2.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional: For advanced features
matplotlib>=3.5.0  # For charts/visualization
plotly>=5.0.0     # For interactive charts
jupyter>=1.0.0    # For analysis notebooks

# Database (optional, for trade history)
sqlalchemy>=1.4.0
aiosqlite>=0.17.0

# Caching (optional)
redis>=4.3.0
aioredis>=2.0.0
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0
ccxt>=4.0.0
web3>=6.0.0

# Google Cloud Web3 Integration
google-cloud-bigquery==3.13.0     # BigQuery for analytics
google-cloud-pubsub==2.18.0       # Real-time data streaming
google-cloud-firestore==2.13.0    # NoSQL database for cache
google-cloud-kms==2.20.0         # Key management
google-cloud-monitoring==2.16.0  # Monitoring and alerting
google-cloud-logging==3.8.0      # Logging
google-cloud-aiplatform==1.38.0  # Vertex AI for ML
google-cloud-functions==1.14.0   # Cloud Functions for serverless
google-cloud-run==0.10.0         # Cloud Run for containerized services
google-cloud-build==3.18.0       # Cloud Build for CI/CD
google-cloud-storage==2.10.0     # Cloud Storage for data
google-auth==2.23.0              # Google Cloud authentication
google-cloud-artifact-registry==1.9.0  # Artifact Registry for containers